pyproject.toml
[project]
name = "11labs-hackathon-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=1.51.0",
    "ipykernel>=6.29.5",
    "mistral>=19.0.0",
    "mistralai>=1.5.0",
    "openai>=1.63.2",
    "pandas>=2.2.3",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "setuptools-rust>=1.10.2",
    "streamlit>=1.42.2",
]
